pyproject.toml
[tool.poetry]
name = "Devoirs24Project"
version = "0.1.0"
description = "devoirs"
authors = ["Annabse27 <annabse27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
django-filter = "^24.3"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
drf-spectacular = "^0.27.2"
stripe = "^10.12.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
